requirements.txt
# =====================================================
# OpenMetadata Ingestion Requirements
# =====================================================
# Core OpenMetadata SDK
openmetadata-ingestion==1.9.7

# AWS SDK for S3 integration
boto3>=1.35.0
botocore>=1.35.0
s3transfer>=0.14.0

# Data processing and analysis
pandas>=1.5.0
numpy>=1.21.0

# Sample data generation
faker>=19.0.0

# Data format support
pyarrow>=12.0.0  # For Parquet files
fastparquet>=0.8.0  # Alternative Parquet engine

# HTTP requests
requests>=2.28.0
urllib3>=1.26.0

# Configuration and environment
python-dotenv>=0.19.0
PyYAML>=6.0

# Date/time handling
python-dateutil>=2.8.0

# Logging and utilities
colorama>=0.4.4
rich>=12.0.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Type hints
types-PyYAML>=6.0.0
types-requests>=2.28.0

# Additional data connectors (if needed)
# Uncomment based on your specific data sources:
# sqlalchemy>=1.4.0  # For database connections
# psycopg2-binary>=2.9.0  # For PostgreSQL
# pymongo>=4.0.0  # For MongoDB
# confluent-kafka>=1.9.0  # For Kafka
# elasticsearch>=8.0.0  # For Elasticsearch

# Development tools (optional)
# black>=22.0.0  # Code formatting
# flake8>=5.0.0  # Linting
# mypy>=0.991  # Type checking
